feat: add character appearances persistence mechanism
When a query is made, the application's very first action is to look for the up-to-date data on the queried character's appearances. It will only rely on the external API in case the data isn't recorded in the database or if the available data is considered outdated. It was established that the available data must be updated in case the last update on the character version's appearances was made over 3 days prior to the query

+++ src/main/java/com/santos/ravenapi/search/client/FandomApiClient.java
@@ -1,8 +1,8 @@
 package com.santos.ravenapi.search.client;
 
-import com.santos.ravenapi.model.dto.search.appearances.FandomAppearancesDTO;
-import com.santos.ravenapi.model.dto.search.disambiguation.FandomDisambiguationDTO;
-import com.santos.ravenapi.model.dto.search.issues.FandomIssueDetailsDTO;
+import com.santos.ravenapi.model.dto.appearances.FandomAppearancesDTO;
+import com.santos.ravenapi.model.dto.disambiguation.FandomDisambiguationDTO;
+import com.santos.ravenapi.model.dto.issues.FandomIssueDetailsDTO;
 import com.santos.ravenapi.search.enums.PublisherEndpointEnum;
 
 public interface FandomApiClient {
@@ -11,7 +11,7 @@ public interface FandomApiClient {
 	
 	FandomAppearancesDTO queryAppearances(PublisherEndpointEnum endpoint, String character, String cont);
 	
-	FandomIssueDetailsDTO queryIssueDetails(PublisherEndpointEnum endpoint, int issueId);
+	FandomIssueDetailsDTO queryIssueDetails(PublisherEndpointEnum endpoint, long issueId);
 	
 	FandomDisambiguationDTO queryDisambiguation(PublisherEndpointEnum endpoint, String character);
 }
